+++ shpc/client/install.py
@@ -16,4 +16,9 @@ def main(args, parser, extra, subparser):
         module=args.module,
         container_tech=args.container_tech,
     )
+
+    # Update config settings on the fly
+    cli.settings.update_params(args.config_params)
+
+    # And do the install
     cli.install(args.install_recipe, symlink=args.symlink)