+++ shpc/client/listing.py
@@ -14,4 +14,6 @@ def main(args, parser, extra, subparser):
         container_tech=args.container_tech,
     )
 
+    # Update config settings on the fly
+    cli.settings.update_params(args.config_params)
     cli.list(args.pattern, args.names_only, short=args.short)