+++ shpc/client/test.py
@@ -14,6 +14,8 @@ def main(args, parser, extra, subparser):
         container_tech=args.container_tech,
     )
 
+    # Update config settings on the fly
+    cli.settings.update_params(args.config_params)
     cli.test(
         args.module_name,
         test_exec=args.test_exec,