+++ shpc/client/uninstall.py
@@ -13,4 +13,7 @@ def main(args, parser, extra, subparser):
         module=args.module,
         container_tech=args.container_tech,
     )
+
+    # Update config settings on the fly
+    cli.settings.update_params(args.config_params)
     cli.uninstall(args.uninstall_recipe, args.force)