+++ shpc/main/schemas.py
@@ -140,6 +140,7 @@ settingsProperties = {
     "default_version": {"type": "boolean"},
     "enable_tty": {"type": "boolean"},
     "symlink_home": {"type": ["string", "null"]},
+    "symlink_tree": {"type": "boolean"},
     "wrapper_scripts": wrapper_scripts,
     "container_tech": {"type": "string", "enum": ["singularity", "podman", "docker"]},
     "singularity_shell": {"type": "string", "enum": shells},